Moving notification permission logic to initialization, instead of SplashScreen
